% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_codes.R, R/landings.R
\name{country_codes}
\alias{country_codes}
\alias{landings}
\title{country_codes}
\usage{
country_codes()

landings()
}
\description{
Function has been deprecated. Now replaced by \code{of_country_codes}
}

