\name{print.rfisheries}
\alias{print.rfisheries}
\title{Print a summary for an rfisheries object}
\usage{
  \method{print}{rfisheries} (x, ...)
}
\arguments{
  \item{x}{An object of class \code{Altmetric}}

  \item{...}{additional arguments}
}
\description{
  Print a summary for an rfisheries object
}

