% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfisheries_package.R
\name{rfisheries-deprecated}
\alias{rfisheries-deprecated}
\title{Deprecated functions in rfisheries}
\description{
\itemize{
 \item \code{\link{country_codes}}: service no longer provided
 \item \code{\link{species_codes}}: service no longer provided
 \item \code{\link{landings}}: service no longer provided
}
}

