% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfisheries_package.R
\docType{data}
\name{country_code_data}
\alias{country_code_data}
\title{country_code_data}
\format{A data table with 239 rows and 2 variables}
\usage{
data(country_code_data)
}
\description{
A dataset containing full list of Country Codes. Use this dataset if loading from the web is slow.
}
\keyword{datasets}

