\docType{data}
\name{species_code_data}
\alias{species_code_data}
\title{species_code_data}
\format{A data table with 11562 rows and 5 variables}
\usage{
data(species_code_data)
}
\description{
A dataset containing full list of Species Codes. Use this
dataset if loading from the web is slow.
}
\keyword{datasets}

